refactor: cut over wallet entrypoints
